improved code quality
