+++ app/posts/[id]/page.tsx
@@ -12,7 +12,6 @@ import { inter } from "@/custom-fonts/fonts";
 import { useState, useEffect } from "react";
 import { Turn as Hamburger } from "hamburger-react";
 
-import Skeleton, { SkeletonTheme } from "react-loading-skeleton";
 import "react-loading-skeleton/dist/skeleton.css";
 import { Discussion, Post, Reply } from "@/types/post";
 import { notFound } from "next/navigation";
@@ -25,6 +24,9 @@ import { useRouter } from "next/navigation";
 import { MdEdit, MdOutlineChat } from "react-icons/md";
 import { State } from "@/types/user";
 import { LuLogIn } from "react-icons/lu";
+import Headings from "@/app/components/Posts/Headings";
+import Tags from "@/app/components/Posts/Tags";
+import LoadingSkeleton from "@/app/components/Posts/LoadingSkeleton";
 
 type Params = {
   params: { id: string };
@@ -103,13 +105,6 @@ export default function Post({ params }: Params) {
   const [inputValueStatus, setInputValueStatus] = useState<
     string | { discussion: string; reply: string }
   >("");
-
-  let currentId = 0;
-
-  const getId = () => {
-    currentId += 1;
-    return currentId;
-  };
 
   function handleInput(type: "discussion" | "reply", replyId?: string) {
     if (type == "discussion") {
@@ -258,50 +253,7 @@ export default function Post({ params }: Params) {
   const [openMenu, setOpenMenu] = useState<boolean>(false);
 
   return loading ? (
-    <div className="h-screen bg-slate-200 dark:bg-slate-800">
-      <SkeletonTheme
-        baseColor={theme == "dark" ? "#333344" : "#aeaeae"}
-        highlightColor={theme == "dark" ? "#aaa" : "#999"}
-        duration={0.7}
-      >
-        <div className="mx-auto mb-3 w-[95%] pt-5">
-          <Skeleton height={100} />
-        </div>
-        <div className="sm:flex">
-          <div className="h-0 w-0 md:h-fit md:w-[30vw] md:pl-3">
-            <div className="mb-1 w-[80%]">
-              <Skeleton height={40} />
-            </div>
-            <div className="mb-1 w-[70%]">
-              <Skeleton height={40} />
-            </div>
-            <div className="mb-1 w-[90%]">
-              <Skeleton height={40} />
-            </div>
-            <div className="mb-1 w-[65%]">
-              <Skeleton height={40} />
-            </div>
-          </div>
-          <div className="sm:w-[80vw] sm:pr-3 md:w-[50vw]">
-            <div className="mb-2 ml-3 mt-5 w-[90%]">
-              <Skeleton height={80} />
-            </div>
-            <div className="ml-3 w-[88%]">
-              <Skeleton height={170} />
-            </div>
-            <div className="mb-2 ml-3 mt-5 w-[70%]">
-              <Skeleton height={80} />
-            </div>
-            <div className="ml-3 w-[88%]">
-              <Skeleton height={130} />
-            </div>
-          </div>
-          <div className="w-0 sm:w-[20vw] sm:pr-3">
-            <Skeleton height={300} />
-          </div>
-        </div>
-      </SkeletonTheme>
-    </div>
+    <LoadingSkeleton />
   ) : (
     <div className="dark:bg-slate-900 dark:text-slate-100">
       <Toaster position="top-center" />
@@ -387,39 +339,7 @@ export default function Post({ params }: Params) {
         )}
         <div className="relative">
           {/* Headings */}
-          <ul
-            className={
-              "fixed left-0 top-[50px] z-[100] h-[100vh] w-[50vw] overflow-y-auto rounded-md bg-slate-800 px-3 py-2 text-slate-400 transition-transform duration-200 md:sticky md:left-2 md:top-10 md:mx-2 md:mr-2 md:inline-flex md:h-fit md:max-h-[70vh] md:w-[15vw] md:flex-col md:justify-start md:bg-[rgba(250,250,250,0.65)] md:p-2 md:text-inherit md:text-slate-600 md:dark:bg-[rgba(50,50,50,0.65)]" +
-              (openMenu
-                ? " translate-x-0"
-                : " -translate-x-[80vw] md:translate-x-0")
-            }
-          >
-            {headings.map((h) => {
-              return (
-                <li
-                  key={h}
-                  className="my-4 w-fit text-lg underline-offset-2 hover:underline dark:text-slate-300 md:my-2 md:text-sm"
-                >
-                  <p
-                    className={`${getId()} cursor-pointer`}
-                    onClick={(e) => {
-                      const id = e.currentTarget.classList[0];
-                      const heading = document.getElementById(id);
-                      if (heading != null) {
-                        heading.scrollIntoView({
-                          behavior: "smooth",
-                          block: "start",
-                        });
-                      }
-                    }}
-                  >
-                    {h}
-                  </p>
-                </li>
-              );
-            })}
-          </ul>
+          <Headings headings={headings} openMenu={openMenu} />
 
           <div className="sm:inline-flex md:max-w-[75vw]">
             <div className="flex-[3]">
@@ -432,22 +352,7 @@ export default function Post({ params }: Params) {
                 {postData && <ShowMarkdown data={postData.content} />}
               </div>
             </div>
-
-            <div className="my-3 flex flex-[1] flex-wrap items-center justify-center text-center text-slate-200 sm:flex-col sm:justify-start">
-              <p className="mb-2 hidden text-lg text-slate-800 dark:text-slate-200 sm:block">
-                Tags
-              </p>
-              {postData?.tags.map((tag) => {
-                return (
-                  <div
-                    key={tag}
-                    className="mx-[2px] my-[2px] rounded-xl bg-slate-600 px-[4px] py-[2px] dark:text-slate-300 sm:w-[80%]"
-                  >
-                    {tag}
-                  </div>
-                );
-              })}
-            </div>
+            <Tags tags={postData?.tags} />
           </div>
         </div>
         <div className={inter.className}>
@@ -468,9 +373,7 @@ export default function Post({ params }: Params) {
               uploadMarkdown={submitData}
               discussionHandler={setTakeNewMarkdownInput}
             />
-          ) : (
-            ""
-          )}
+          ) : null}
         </div>
       </div>
     </div>
